accessible/tests/mochitest/text/test_general.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>

<window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        title="Tests: XUL label text interface">

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js" />
  <script src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>

  <script type="application/javascript"
          src="../common.js"></script>
  <script type="application/javascript"
          src="../role.js"></script>
  <script type="application/javascript"
          src="../text.js"></script>

  <script type="application/javascript">
  <![CDATA[
    ////////////////////////////////////////////////////////////////////////////
    // Testing

    var gQueue = null;
    function doTests()
    {
      //////////////////////////////////////////////////////////////////////////
      // XUL label

      var ids = ["label1", "label2"];

      testCharacterCount(ids, 5);

      testText(ids, 0, -1, "Hello");
      testText(ids, 0, 1, "H");

      testCharAfterOffset(ids, 0, "e", 1, 2);
      testCharBeforeOffset(ids, 1, "H", 0, 1);
      testCharAtOffset(ids, 1, "e", 1, 2);

      //////////////////////////////////////////////////////////////////////////
      // XUL textbox

      testTextAtOffset([ getNode("tbox1").inputField ], BOUNDARY_LINE_START,
                       [ [ 0, 4, "test", 0, 4 ] ]);

      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTests);
  ]]>
  </script>

  <vbox flex="1" style="overflow: auto;">
  <body xmlns="http://www.w3.org/1999/xhtml">
    <a target="_blank"
        href="https://bugzilla.mozilla.org/show_bug.cgi?id=396166"
        title="xul:label@value accessible should implement nsIAccessibleText">
      Bug 396166
    </a>
    <a target="_blank"
        href="https://bugzilla.mozilla.org/show_bug.cgi?id=899433"
        title="Accessibility returns empty line for last line in certain cases">
      Bug 899433
    </a>
    <p id="display"></p>
    <div id="content" style="display: none">
    </div>
    <pre id="test">
    </pre>
  </body>
  <label id="label1" value="Hello"/>
  <label id="label2">Hello</label>

  <textbox id="tbox1" value="test"/>
  </vbox>
</window>
